agents/procedure_agent/requirements.txt
# FastAPI web framework
fastapi==0.115.14
pydantic==2.11.7
pydantic_core==2.33.2
starlette==0.46.2
typing_extensions==4.14.1
requests 
# Uvicorn server
uvicorn==0.35.0
h11==0.16.0
click==8.2.1

# For reading .env file
python-dotenv==1.1.1

# Snowflake Database Connector and its dependencies
snowflake-connector-python==3.16.0
# ... (pip will automatically pull in snowflake's dependencies like cryptography, cffi, etc.)